New helper for analysis app and email reporting
Move all reporting functionality to a new helper module, purely for tidiness, reducing the length of the OpencBot::CompanyFetcherBot module.
